signal and wait have been changed at the phaser-level.
This change imposes two restrictions:
- Only signaller tasks can perform a signal;
- Similarly, only waiter tasks can perform a wait.

Historically, this constraint was not being modelled as we wanted to
define wait_all and signal_all. The concepts are now clearer since
for the definition of wait_all and signal_all we now use try_wait
and try_signal, respectively. Progress and subject rediction were
updated.